requirements.txt
backports.cached-property==1.0.2
greenlet==1.1.2
grpcio==1.47.0
grpcio-tools==1.47.0
natsort==8.2.0
overrides==6.2.0
playwright==1.25.2
protobuf==3.20.1
pyee==8.1.0
PyYAML==6.0
robotframework==5.0.1
robotframework-assertion-engine==0.5.1
robotframework-browser==13.4.0
robotframework-pabot==2.8.0
robotframework-playwright==0.1.0
robotframework-pythonlibcore==3.0.0
robotframework-stacktrace==0.4.1
six==1.16.0
typing_extensions==4.3.0
websockets==10.1
wrapt==1.14.1
